Analyze Data Flow: Generalize "Group by leaf expression nullness" action
