+++ java/java-impl/src/com/intellij/slicer/CanItBeNullAction.java
@@ -15,33 +15,18 @@
  */
 package com.intellij.slicer;
 
-import com.intellij.icons.AllIcons;
-import com.intellij.openapi.actionSystem.AnAction;
-import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.psi.*;
 import org.jetbrains.annotations.NotNull;
 
 import javax.swing.tree.DefaultMutableTreeNode;
 
-class CanItBeNullAction extends AnAction {
-  private final SliceTreeBuilder myTreeBuilder;
-  private static final String TEXT = "Group by leaf expression nullness";
-
+class CanItBeNullAction extends GroupByNullnessActionBase {
   CanItBeNullAction(@NotNull SliceTreeBuilder treeBuilder) {
-    super(TEXT, "Determine whether null can flow into this expression", AllIcons.Debugger.Db_disabled_breakpoint_process);
-    myTreeBuilder = treeBuilder;
+    super(treeBuilder);
   }
 
   @Override
-  public void update(AnActionEvent e) {
-    e.getPresentation().setText(TEXT + (myTreeBuilder.analysisInProgress ? " (Analysis in progress)" : ""));
-    e.getPresentation().setEnabled(isAvailable());
-  }
-
-  private boolean isAvailable() {
-    if (myTreeBuilder.analysisInProgress) return false;
-    if (!myTreeBuilder.dataFlowToThis) return false;
-    if (myTreeBuilder.splitByLeafExpressions) return false;
+  protected boolean isAvailable() {
     DefaultMutableTreeNode root = myTreeBuilder.getRootNode();
     if (root == null) return false;
     SliceRootNode rootNode = (SliceRootNode)root.getUserObject();
@@ -57,10 +42,5 @@ class CanItBeNullAction extends AnAction {
       type = null;
     }
     return type instanceof PsiClassType || type instanceof PsiArrayType;
-  }
-
-  @Override
-  public void actionPerformed(AnActionEvent e) {
-    myTreeBuilder.switchToLeafNulls();
   }
 }
